Move searchbar to its own component
